+++ glue/src/stateful/mod.rs
@@ -143,13 +143,12 @@ pub struct Input<Upstream, Provider> {
 /// the block tree and applying changesets to the underlying databases on
 /// finalization.
 ///
-/// Stateful runs each verification, including any lazy [`apply`](Self::apply)
-/// calls it requires, on a disposable clone of the application. Multiple
-/// [`verify`](Self::verify) and `apply` calls may overlap each other and may
-/// overlap [`propose`](Self::propose) or [`finalized`](Self::finalized).
-/// Equivalent execution results may be reused across requests, so correctness
-/// must not depend on every clone observing every call. State whose visibility
-/// affects correctness across calls must be shared explicitly.
+/// [`Stateful`] may freely clone the application and invoke its methods
+/// concurrently. Implementors should treat `Application` as a stateless,
+/// deterministic state machine: given the same method inputs and database
+/// state, every clone must produce the same state-transition result. Mutable
+/// state that affects those results must live in the provided database batches
+/// or database set.
 pub trait Application<E>: Clone + Send + 'static
 where
     E: Rng + Spawner + Metrics + Clock,
